Dashboard logo added , mainlayout setup
