+++ src/app/dashboard/layout.js
@@ -1,8 +1,12 @@
 "use client";
 
 import Sidebar from "@/components/layout/Sidebar";
+import Image from "next/image";
 import { useState } from "react";
 import { FaBars } from "react-icons/fa";
+import Logo from "../../../public/images/Logo.png";
+import Link from "next/link";
+
 
 export default function DashboardLayout({ children }) {
   const [isSidebarOpen, setIsSidebarOpen] = useState(false);
@@ -19,19 +23,23 @@ export default function DashboardLayout({ children }) {
           className="lg:hidden flex items-center justify-between px-4 py-3 
                         bg-white dark:bg-gray-900 shadow-md border-b dark:border-gray-700"
         >
+          <Link 
+          href="/"
+          aria-label="Company"
+          title="Company"
+          className="text-lg font-bold text-gray-700 dark:text-gray-200 font-gabriela flex items-center gap-2 ">
+             <Image src={Logo} alt="logo" className="w-10 md:w-12" />  <p className="text-xl">Eventoria</p>
+          </Link>
           <button
             onClick={() => setIsSidebarOpen(true)}
             className="text-gray-800 dark:text-gray-200 text-xl"
           >
             <FaBars />
           </button>
-          <h1 className="text-lg font-bold text-gray-700 dark:text-gray-200">
-            Dashboard
-          </h1>
         </div>
 
         {/* Page Content */}
-        <main className="flex-1 p-4 text-gray-900 dark:text-gray-100 overflow-y-auto">
+        <main className="flex-1 p-4 text-gray-900 dark:text-gray-100 overflow-y-auto font-gabriela">
           {children}
         </main>
       </div>
