+++ src/app/layout.js
@@ -1,0 +1,36 @@
+
+
+import { Gabriela, Poppins } from "next/font/google";
+import "./globals.css";
+import AppProviders from "@/components/layout/AppProviders"; 
+
+
+const poppins = Poppins({
+ variable: "--font-poppins",
+ subsets: ["latin"],
+ weight: ["400", "500", "600", "700", "800", "900"],
+});
+
+export const metadata = { 
+ title: "Eventoria",
+ description: "Generated by create next app",
+};
+
+const gabriela = Gabriela({
+ variable: "--font-gabriela",
+ subsets: ["latin"],
+ weight: "400",
+});
+
+export default function RootLayout({ children }) {
+ return (
+  <html lang="en" suppressHydrationWarning> 
+   <body className={`${gabriela.variable} ${poppins.variable} antialiased`}>
+  
+    <AppProviders>
+            {children}
+        </AppProviders>
+   </body>
+  </html>
+ );
+}
